+++ app/storage/migrations/20260919152201_backfill_public_faq_documentation_tag/migration.sql
@@ -1,0 +1,4 @@
+UPDATE "Documentation"
+SET tags = ARRAY['public_faq']::"DocumentationTag"[]
+WHERE "isFaq" = true
+  AND tags = '{}';